🐛 Fixed drop mortar shell not firing, make added_charge_power field explicit in examples, cannon pitch fixes, mount info tooltip formatting

+++ common/src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/AbstractBigCannonProjectile.java
@@ -79,7 +79,7 @@ public abstract class AbstractBigCannonProjectile<T extends BigCannonProjectileP
 		return new CannonDamageSource(CreateBigCannons.MOD_ID + ".big_cannon_projectile", this, null);
 	}
 
-	public int addedChargePower() {
+	public float addedChargePower() {
 		T properties = this.getProperties();
 		return properties == null ? 0 : properties.addedChargePower();
 	}
